+++ static/templates/dialog_widget.hbs
