+++ static/templates/dialog_widget_heading.hbs
